Move CFB, OFB, and CFB8 modes to decrepit module
These cipher modes are deprecated and moved to the decrepit module. They
are re-exported from the original modes module with deprecation warnings.

- Added DeprecatedIn47 alias to utils module
- Created new decrepit/ciphers/modes.py with CFB, OFB, and CFB8 classes
- Re-exported modes from primitives/ciphers/modes.py with deprecation warnings
- Updated CHANGELOG.rst following the pattern used for CAST5 deprecation
- Added comprehensive documentation in docs/hazmat/decrepit/modes.rst

These modes will be removed from primitives.ciphers.modes in 49.0.0.
