+++ src/cryptography/hazmat/primitives/ciphers/modes.py
@@ -4,106 +4,33 @@
 
 from __future__ import annotations
 
-import abc
-
 from cryptography import utils
 from cryptography.exceptions import UnsupportedAlgorithm, _Reasons
 from cryptography.hazmat.primitives._cipheralgorithm import (
     BlockCipherAlgorithm,
     CipherAlgorithm,
 )
+from cryptography.hazmat.primitives._modes import (
+    Mode as Mode,
+)
+from cryptography.hazmat.primitives._modes import (
+    ModeWithAuthenticationTag as ModeWithAuthenticationTag,
+)
+from cryptography.hazmat.primitives._modes import (
+    ModeWithInitializationVector as ModeWithInitializationVector,
+)
+from cryptography.hazmat.primitives._modes import (
+    ModeWithNonce as ModeWithNonce,
+)
+from cryptography.hazmat.primitives._modes import (
+    ModeWithTweak as ModeWithTweak,
+)
+from cryptography.hazmat.primitives._modes import (
+    _check_aes_key_length,
+    _check_iv_and_key_length,
+    _check_nonce_length,
+)
 from cryptography.hazmat.primitives.ciphers import algorithms
-
-
-class Mode(metaclass=abc.ABCMeta):
-    @property
-    @abc.abstractmethod
-    def name(self) -> str:
-        """
-        A string naming this mode (e.g. "ECB", "CBC").
-        """
-
-    @abc.abstractmethod
-    def validate_for_algorithm(self, algorithm: CipherAlgorithm) -> None:
-        """
-        Checks that all the necessary invariants of this (mode, algorithm)
-        combination are met.
-        """
-
-
-class ModeWithInitializationVector(Mode, metaclass=abc.ABCMeta):
-    @property
-    @abc.abstractmethod
-    def initialization_vector(self) -> utils.Buffer:
-        """
-        The value of the initialization vector for this mode as bytes.
-        """
-
-
-class ModeWithTweak(Mode, metaclass=abc.ABCMeta):
-    @property
-    @abc.abstractmethod
-    def tweak(self) -> utils.Buffer:
-        """
-        The value of the tweak for this mode as bytes.
-        """
-
-
-class ModeWithNonce(Mode, metaclass=abc.ABCMeta):
-    @property
-    @abc.abstractmethod
-    def nonce(self) -> utils.Buffer:
-        """
-        The value of the nonce for this mode as bytes.
-        """
-
-
-class ModeWithAuthenticationTag(Mode, metaclass=abc.ABCMeta):
-    @property
-    @abc.abstractmethod
-    def tag(self) -> bytes | None:
-        """
-        The value of the tag supplied to the constructor of this mode.
-        """
-
-
-def _check_aes_key_length(self: Mode, algorithm: CipherAlgorithm) -> None:
-    if algorithm.key_size > 256 and algorithm.name == "AES":
-        raise ValueError(
-            "Only 128, 192, and 256 bit keys are allowed for this AES mode"
-        )
-
-
-def _check_iv_length(
-    self: ModeWithInitializationVector, algorithm: BlockCipherAlgorithm
-) -> None:
-    iv_len = len(self.initialization_vector)
-    if iv_len * 8 != algorithm.block_size:
-        raise ValueError(f"Invalid IV size ({iv_len}) for {self.name}.")
-
-
-def _check_nonce_length(
-    nonce: utils.Buffer, name: str, algorithm: CipherAlgorithm
-) -> None:
-    if not isinstance(algorithm, BlockCipherAlgorithm):
-        raise UnsupportedAlgorithm(
-            f"{name} requires a block cipher algorithm",
-            _Reasons.UNSUPPORTED_CIPHER,
-        )
-    if len(nonce) * 8 != algorithm.block_size:
-        raise ValueError(f"Invalid nonce size ({len(nonce)}) for {name}.")
-
-
-def _check_iv_and_key_length(
-    self: ModeWithInitializationVector, algorithm: CipherAlgorithm
-) -> None:
-    if not isinstance(algorithm, BlockCipherAlgorithm):
-        raise UnsupportedAlgorithm(
-            f"{self} requires a block cipher algorithm",
-            _Reasons.UNSUPPORTED_CIPHER,
-        )
-    _check_aes_key_length(self, algorithm)
-    _check_iv_length(self, algorithm)
 
 
 class CBC(ModeWithInitializationVector):
@@ -153,48 +80,6 @@ class ECB(Mode):
     name = "ECB"
 
     validate_for_algorithm = _check_aes_key_length
-
-
-class OFB(ModeWithInitializationVector):
-    name = "OFB"
-
-    def __init__(self, initialization_vector: utils.Buffer):
-        utils._check_byteslike("initialization_vector", initialization_vector)
-        self._initialization_vector = initialization_vector
-
-    @property
-    def initialization_vector(self) -> utils.Buffer:
-        return self._initialization_vector
-
-    validate_for_algorithm = _check_iv_and_key_length
-
-
-class CFB(ModeWithInitializationVector):
-    name = "CFB"
-
-    def __init__(self, initialization_vector: utils.Buffer):
-        utils._check_byteslike("initialization_vector", initialization_vector)
-        self._initialization_vector = initialization_vector
-
-    @property
-    def initialization_vector(self) -> utils.Buffer:
-        return self._initialization_vector
-
-    validate_for_algorithm = _check_iv_and_key_length
-
-
-class CFB8(ModeWithInitializationVector):
-    name = "CFB8"
-
-    def __init__(self, initialization_vector: utils.Buffer):
-        utils._check_byteslike("initialization_vector", initialization_vector)
-        self._initialization_vector = initialization_vector
-
-    @property
-    def initialization_vector(self) -> utils.Buffer:
-        return self._initialization_vector
-
-    validate_for_algorithm = _check_iv_and_key_length
 
 
 class CTR(ModeWithNonce):
@@ -266,3 +151,51 @@ class GCM(ModeWithInitializationVector, ModeWithAuthenticationTag):
                 f"Authentication tag cannot be more than {block_size_bytes} "
                 "bytes."
             )
+
+
+# Import decrepit modes after base classes are defined to avoid circular
+# imports
+from cryptography.hazmat.decrepit.ciphers.modes import (  # noqa: E402
+    CFB as CFB,
+)
+from cryptography.hazmat.decrepit.ciphers.modes import (  # noqa: E402
+    CFB8 as CFB8,
+)
+from cryptography.hazmat.decrepit.ciphers.modes import (  # noqa: E402
+    OFB as OFB,
+)
+
+utils.deprecated(
+    OFB,
+    __name__,
+    "OFB has been moved to "
+    "cryptography.hazmat.decrepit.ciphers.modes.OFB and "
+    "will be removed from "
+    "cryptography.hazmat.primitives.ciphers.modes in 49.0.0.",
+    utils.DeprecatedIn47,
+    name="OFB",
+)
+
+
+utils.deprecated(
+    CFB,
+    __name__,
+    "CFB has been moved to "
+    "cryptography.hazmat.decrepit.ciphers.modes.CFB and "
+    "will be removed from "
+    "cryptography.hazmat.primitives.ciphers.modes in 49.0.0.",
+    utils.DeprecatedIn47,
+    name="CFB",
+)
+
+
+utils.deprecated(
+    CFB8,
+    __name__,
+    "CFB8 has been moved to "
+    "cryptography.hazmat.decrepit.ciphers.modes.CFB8 and "
+    "will be removed from "
+    "cryptography.hazmat.primitives.ciphers.modes in 49.0.0.",
+    utils.DeprecatedIn47,
+    name="CFB8",
+)
